+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Version - 0]
+## Fix [v0.6.1] 2022-08-01
+- fixed search by set state page in index farmer, droner, user staff
 ## Feature [v0.6.0] 2022-08-01
 - index droner 
 ## Fix [v0.5.2] 2022-08-01